Dropwizard

+++ CRS-JEDI-JAVA-Development-DropWizard/src/main/java/com/flipkart/service/AdminServiceOperation.java
@@ -1,10 +1,7 @@
 package com.flipkart.service;
 
-import com.flipkart.bean.Course;
-import com.flipkart.bean.Professor;
-import com.flipkart.bean.Student;
-import com.flipkart.dao.AdminDAOImpl;
-import com.flipkart.dao.UserDAOImpl;
+import com.flipkart.bean.*;
+import com.flipkart.dao.*;
 import com.flipkart.exception.*;
 
 import java.util.List;
@@ -16,59 +13,61 @@ public class AdminServiceOperation implements AdminInterface {
     AdminDAOImpl adminDaoImpl=new AdminDAOImpl();
     Scanner sc=new Scanner(System.in);
 
-    public void approveStudentRegistration(int studentID) throws StudentNotFoundForApprovalException {
+    public void approveStudentRegistration(int studentID) throws StudentNotFoundForApprovalException,StudentAlreadyApproved {
+        UserDAO dao = new UserDAOImpl();
+        User u = dao.getUserByID(studentID);
+        if(u==null)
+        {
+            throw new StudentNotFoundForApprovalException(studentID);
+        }
+        if(u.getIsApproved())
+        {
+            throw new StudentAlreadyApproved(studentID);
+        }
         adminDaoImpl.approveStudent(studentID);
     }
     @Override
-    public void addAdmin(int userID, String userName, String password, String role, boolean isApproved, String name){
-
+    public void addAdmin(Admin a) throws DuplicateUserException{
         UserDAOImpl userDAO = new UserDAOImpl();
-
-        userDAO.register(userID, userName, password, role, true);
-
-        adminDaoImpl.addAdmin(userID, name);
+        if(userDAO.getUserByID(a.getUserID())!=null)throw new DuplicateUserException(a.getUserID());
+        userDAO.register(a.getUserID(), a.getName(), a.getPassword(), a.getRole(), true);
+        adminDaoImpl.addAdmin(a.getUserID(), a.getName());
     }
-    public void addProfessor(int userID,String userName,String password,String role,String name,String dept,String designation) throws UserIdAlreadyInUseException, ProfessorNotAddedException {
+    public void addProfessor(Professor p) throws DuplicateUserException, ProfessorNotAddedException {
 
         UserDAOImpl userDAO = new UserDAOImpl();
-
-        userDAO.register(userID, userName, password, role, true);
-
-        Professor professor = new Professor();
-        professor.setUserID(userID);
-        professor.setName(name);
-        professor.setDepartment(dept);
-        professor.setDesignation(designation);
-
-        adminDaoImpl.addProfessor(professor);
+        if(userDAO.getUserByID(p.getUserID())!=null)throw new DuplicateUserException(p.getUserID());
+        userDAO.register(p.getUserID(), p.getName(), p.getPassword(), p.getRole(), true);
+        try {
+            adminDaoImpl.addProfessor(p);
+        }catch (ProfessorNotAddedException e)
+        {
+            throw new ProfessorNotAddedException((p.getUserID()));
+        }
         System.out.println("Professor added successfully.");
     }
 
-    public void addCourse(int courseID , String courseName, int semID) {
+    public void addCourse(int courseID , String courseName, int semID) throws CourseAlreadyPresentException {
         Course c1=new Course();
         c1.setCourseID(courseID);
         c1.setCourseName(courseName);
-        try {
-            adminDaoImpl.addCourse(c1, semID);
-        }
-        catch(CourseAlreadyPresentException e)
-        {
-            System.out.println(e.getMessage());
-            return;
-        }
-        System.out.println(courseName + " added successfully.");
+        ProferssorDAO dao =new ProfessorDAOImpl();
+        Course c = dao.getCourseByID(courseID);
+        if(c!=null)throw new CourseAlreadyPresentException(courseID);
+        adminDaoImpl.addCourse(c1, semID);
 
     }
 
     public void removeCourse(int semId , int courseId) throws CourseNotDeletedException, CourseNotFoundException {
+        ProferssorDAO dao = new ProfessorDAOImpl();
+        Course c = dao.getCourseByID(courseId);
+        if(c==null)throw new CourseNotFoundException(courseId);
         try {
             adminDaoImpl.deleteCourse(courseId);
         }
-
         catch(CourseNotDeletedException e)
         {
-            System.out.println(e.getMessage());
-            return;
+            throw new CourseNotDeletedException(courseId);
         }
 
     }
